# spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Puja Booking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/offers": {
      "get": {
        "tags": [
          "offers"
        ],
        "summary": "List Offers",
        "description": "Live offers \u2014 area label only; no customer address or phone (\u00a721.4).",
        "operationId": "list_offers_v1_offers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/offers/{assignment_id}/accept": {
      "post": {
        "tags": [
          "offers"
        ],
        "summary": "Accept",
        "operationId": "accept_v1_offers__assignment_id__accept_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/offers/{assignment_id}/reject": {
      "post": {
        "tags": [
          "offers"
        ],
        "summary": "Reject",
        "operationId": "reject_v1_offers__assignment_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OfferCard": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignment Id"
          },
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "puja_name": {
            "type": "string",
            "title": "Puja Name"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "scheduled_time": {
            "type": "string",
            "format": "time",
            "title": "Scheduled Time"
          },
          "area_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Label"
          },
          "payment_mode": {
            "type": "string",
            "title": "Payment Mode"
          },
          "total_amount": {
            "type": "string",
            "title": "Total Amount"
          },
          "amount_due_offline": {
            "type": "string",
            "title": "Amount Due Offline"
          },
          "booking_class": {
            "type": "string",
            "title": "Booking Class"
          },
          "urgency": {
            "type": "string",
            "title": "Urgency"
          },
          "urgency_escalated": {
            "type": "boolean",
            "title": "Urgency Escalated"
          }
        },
        "type": "object",
        "required": [
          "assignment_id",
          "booking_id",
          "expires_at",
          "puja_name",
          "scheduled_date",
          "scheduled_time",
          "area_label",
          "payment_mode",
          "total_amount",
          "amount_due_offline",
          "booking_class",
          "urgency",
          "urgency_escalated"
        ],
        "title": "OfferCard"
      },
      "OfferListResponse": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/OfferCard"
            },
            "type": "array",
            "title": "Offers"
          }
        },
        "type": "object",
        "required": [
          "offers"
        ],
        "title": "OfferListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}